requirements.txt
gunicorn==19.4.5
Cython==0.22.1
decorator==4.0.9
numpy==1.9.2
Flask==0.10.1
Flask-Cors==2.1.0
Flask-Uploads==0.2.0
functools32==3.2.3.post2
fuzzywuzzy==0.6.1
awscli==1.7.44
futures==3.0.5 
tinys3==0.1.11
gnureadline==6.3.3

httplib2==0.9.1
httpretty==0.8.6
image==1.4.0

itsdangerous==0.24
Jinja2==2.8
jmespath==0.7.1
jsonrpclib==0.1.3
jsonschema==2.5.1
macholib==1.5.1
MarkupSafe==0.23

oauth2client==1.5.0

Pillow==3.1.1


python-dateutil==2.2
pytz==2013.7
pyzmq==14.7.0
regex==2015.10.5
requests==2.9.1
tornado==4.2.1
Unidecode==0.4.18
virtualenv==13.1.2
Werkzeug==0.11.4

